# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Local File Manager API
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusStatus =
  (typeof HealthStatusStatus)[keyof typeof HealthStatusStatus];

export const HealthStatusStatus = {
  ok: "ok",
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
  uptime: number;
}

export interface ErrorResponse {
  error: string;
}

export interface OkResponse {
  ok: boolean;
}

export type FileEntryType = (typeof FileEntryType)[keyof typeof FileEntryType];

export const FileEntryType = {
  file: "file",
  directory: "directory",
  symlink: "symlink",
  other: "other",
} as const;

export interface FileEntry {
  name: string;
  path: string;
  type: FileEntryType;
  size: number | null;
  permissions: string;
  modifiedAt: string;
}

export interface DirectoryListing {
  path: string;
  parentPath?: string | null;
  entries: FileEntry[];
}

export interface FileContent {
  path: string;
  content: string;
  isBinary: boolean;
  size: number;
}

export interface WriteFileBody {
  path: string;
  content: string;
}

export interface MkdirBody {
  path: string;
}

export interface RenameFileBody {
  oldPath: string;
  newPath: string;
}

export interface ExecCommandBody {
  command: string;
  cwd?: string | null;
}

export interface ExecResult {
  stdout: string;
  stderr: string;
  exitCode: number;
}

export type ListFilesParams = {
  path?: string;
};

export type ReadFileParams = {
  path: string;
};

export type DeleteFileParams = {
  path: string;
  recursive?: boolean;
};