# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Local File Manager API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.enum(["ok"]),
  uptime: zod.number(),
});

/**
 * @summary List directory contents
 */
export const listFilesQueryPathDefault = `/`;

export const ListFilesQueryParams = zod.object({
  path: zod.coerce.string().default(listFilesQueryPathDefault),
});

export const ListFilesResponse = zod.object({
  path: zod.string(),
  parentPath: zod.string().nullish(),
  entries: zod.array(
    zod.object({
      name: zod.string(),
      path: zod.string(),
      type: zod.enum(["file", "directory", "symlink", "other"]),
      size: zod.number().nullable(),
      permissions: zod.string(),
      modifiedAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Read file content
 */
export const ReadFileQueryParams = zod.object({
  path: zod.coerce.string(),
});

export const ReadFileResponse = zod.object({
  path: zod.string(),
  content: zod.string(),
  isBinary: zod.boolean(),
  size: zod.number(),
});

/**
 * @summary Write file content (creates or overwrites)
 */
export const WriteFileBody = zod.object({
  path: zod.string(),
  content: zod.string(),
});

export const WriteFileResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Delete a file or directory
 */
export const deleteFileQueryRecursiveDefault = false;

export const DeleteFileQueryParams = zod.object({
  path: zod.coerce.string(),
  recursive: zod.coerce.boolean().default(deleteFileQueryRecursiveDefault),
});

export const DeleteFileResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Create a directory
 */
export const CreateDirectoryBody = zod.object({
  path: zod.string(),
});

export const CreateDirectoryResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Rename or move a file or directory
 */
export const RenameFileBody = zod.object({
  oldPath: zod.string(),
  newPath: zod.string(),
});

export const RenameFileResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Execute a shell command
 */
export const ExecCommandBody = zod.object({
  command: zod.string(),
  cwd: zod.string().nullish(),
});

export const ExecCommandResponse = zod.object({
  stdout: zod.string(),
  stderr: zod.string(),
  exitCode: zod.number(),
});