# lib/api-zod/src/generated/types/directoryListing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Local File Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { FileEntry } from "./fileEntry";

export interface DirectoryListing {
  path: string;
  parentPath?: string | null;
  entries: FileEntry[];
}