# lib/api-zod/src/generated/types/execCommandBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Local File Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecCommandBody {
  command: string;
  cwd?: string | null;
}