# lib/api-zod/src/generated/types/fileEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Local File Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { FileEntryType } from "./fileEntryType";

export interface FileEntry {
  name: string;
  path: string;
  type: FileEntryType;
  size: number | null;
  permissions: string;
  modifiedAt: Date;
}