# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Local File Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { HealthStatusStatus } from "./healthStatusStatus";

export interface HealthStatus {
  status: HealthStatusStatus;
  uptime: number;
}